requirements.txt
Django==4.2.14
psycopg2==2.9.9
djangorestframework==3.15.2